+++ apps/E2E/src/Menu/consts.ts
@@ -1,0 +1,26 @@
+export const HOMEPAGE_MENU_BUTTON = 'Homepage_Menu_Button';
+export const MENU_TESTPAGE = 'Menu_TestPage';
+
+export const MENUTRIGGER_TEST_COMPONENT = 'MenuTrigger_Test_Component';
+export const MENUPOPOVER_TEST_COMPONENT = 'MenuPopover_Test_Component';
+
+/* Test MenuItem 1 */
+export const MENUITEM_TEST_COMPONENT = 'MenuItem_Test_Component'; // A component on each specific test page
+export const MENUITEM_ACCESSIBILITY_LABEL = 'E2E testing MenuItem accessibility label';
+
+/* Test MenuItem 2 */
+export const MENUITEM_NO_A11Y_LABEL_COMPONENT = 'MenuItem_No_A11y_label_Component';
+export const MENUITEM_TEST_LABEL = 'Test MenuItem2 - No Accessibility Label'; // A component on each specific test page
+
+export const MENU_ON_OPEN = 'Menu_On_Open'; // For testing the open functionality of Menu
+export const MENU_ON_CLOSE = 'Menu_On_Close'; // For testing the open functionality of Menu
+
+/* Test MenuItem 3 */
+export const MENUITEM_DISABLED_COMPONENT = 'MenuItem_Disabled_Component';
+
+/* Test MenuItem 4 */
+export const MENUITEM_FOURTH_COMPONENT = 'MenuItem_Fourth_Component';
+
+export const MENU_CALLBACK_RESET_BUTTON = 'Menu_Reset_Button'; // For resetting the menu item callback text
+
+export const MENUITEM_CALLBACK_LABEL = 'Menu_Item_Callback_Label';