+++ apps/fluent-tester/src/TestComponents/index.ts
@@ -3,6 +3,6 @@ import * as React from 'react';
 export type TestDescription = {
   name: string;
   component: React.FunctionComponent;
-  testPage: string;
+  testPageButton: string;
   platforms: string[];
 };