+++ packages/components/Checkbox/src/CheckboxTokens.ts
@@ -6,7 +6,7 @@ import { CheckboxTokens } from './Checkbox.types';
 export const defaultCheckboxTokens: TokenSettings<CheckboxTokens, Theme> = (t: Theme) =>
   ({
     requiredColor: t.colors.redForeground1,
-    requiredPadding: globalTokens.spacing.s,
+    requiredPadding: globalTokens.size80,
     medium: {
       borderRadius: globalTokens.corner.radius20,
       checkboxBorderWidth: globalTokens.stroke.width10,
@@ -14,15 +14,15 @@ export const defaultCheckboxTokens: TokenSettings<CheckboxTokens, Theme> = (t: T
       checkboxSize: 16,
       checkmarkSize: 12,
       label: {
-        padding: globalTokens.spacing.s,
+        padding: globalTokens.size80,
         borderRadius: globalTokens.corner.radius40,
-        spacingLabelAfter: globalTokens.spacing.m,
+        spacingLabelAfter: globalTokens.size120,
         labelIsBefore: {
-          spacingLabelBefore: globalTokens.spacing.m,
+          spacingLabelBefore: globalTokens.size120,
           spacingLabelAfter: 0,
         },
       },
-      padding: globalTokens.spacing.xs,
+      padding: globalTokens.size40,
       variant: 'bodyStandard',
     },
     large: {
@@ -32,15 +32,15 @@ export const defaultCheckboxTokens: TokenSettings<CheckboxTokens, Theme> = (t: T
       checkboxSize: 20,
       checkmarkSize: 16,
       label: {
-        padding: globalTokens.spacing.s,
+        padding: globalTokens.size80,
         borderRadius: globalTokens.corner.radius40,
-        spacingLabelAfter: globalTokens.spacing.m,
+        spacingLabelAfter: globalTokens.size120,
         labelIsBefore: {
-          spacingLabelBefore: globalTokens.spacing.m,
+          spacingLabelBefore: globalTokens.size120,
           spacingLabelAfter: 0,
         },
       },
-      padding: globalTokens.spacing.xs,
+      padding: globalTokens.size40,
       variant: 'bodyStandard',
     },
     checkboxBackgroundColor: t.colors.menuBackground,