+++ packages/components/ContextualMenu/jest.config.js
@@ -1,2 +1,2 @@
 const { configureReactNativeJest } = require('@fluentui-react-native/scripts');
-module.exports = configureReactNativeJest('ios');
+module.exports = configureReactNativeJest('win32');