+++ packages/components/Menu/src/MenuDivider/MenuDividerTokens.ts
@@ -6,5 +6,5 @@ import { MenuDividerTokens } from './MenuDivider.types';
 export const defaultMenuDividerTokens: TokenSettings<MenuDividerTokens, Theme> = (t: Theme): MenuDividerTokens => ({
   backgroundColor: t.colors.neutralStroke2,
   height: globalTokens.stroke.width10,
-  marginVertical: globalTokens.spacing.xs,
+  marginVertical: globalTokens.size40,
 });