+++ packages/experimental/Link/src/Link.types.ts
@@ -17,6 +17,8 @@ export type LinkState = IPressableState & {
   disabled?: boolean;
   /**
    * Specifies whether the link is inline.
+   * Note: Not supported for win32
+   *
    * @default false
    */
   inline?: boolean;
@@ -48,7 +50,8 @@ export type LinkAppearance = 'default' | 'subtle';
  */
 export interface LinkProps extends IWithPressableOptions<TextProps> {
   /**
-   * link appearance
+   * The appearance of the link, either `default` or `subtle`
+   * @default default
    */
   appearance?: LinkAppearance;
   /**
@@ -61,7 +64,9 @@ export interface LinkProps extends IWithPressableOptions<TextProps> {
    */
   enableFocusRing?: boolean;
   /**
-   * link inline
+   * Whether the link is inline with text
+   * Note: Not supported for win32
+   * @default false
    */
   inline?: boolean;
   /**