+++ packages/experimental/Link/CHANGELOG.md
@@ -1,8 +1,119 @@
 # Change Log - @fluentui-react-native/experimental-link
 
-This log was last generated on Mon, 28 Nov 2022 19:49:35 GMT and should not be manually modified.
+This log was last generated on Tue, 27 Dec 2022 22:21:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.3
+
+Tue, 27 Dec 2022 22:21:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.28
+- Bump @fluentui-react-native/interactive-hooks to v0.22.3
+- Bump @fluentui-react-native/tokens to v0.20.1
+- Bump @fluentui-react-native/text to v0.19.2
+
+## 0.4.2
+
+Tue, 27 Dec 2022 10:18:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.27
+- Bump @fluentui-react-native/interactive-hooks to v0.22.2
+- Bump @fluentui-react-native/text to v0.19.1
+
+## 0.4.1
+
+Wed, 21 Dec 2022 01:18:57 GMT
+
+### Patches
+
+- Fix Link space/enter keyboarding (ruaraki@microsoft.com)
+
+## 0.4.0
+
+Tue, 20 Dec 2022 23:10:17 GMT
+
+### Minor changes
+
+- Force win32 to not have inline links (ruaraki@microsoft.com)
+
+### Patches
+
+- Update link spec with more detail (ruaraki@microsoft.com)
+
+## 0.3.2
+
+Thu, 15 Dec 2022 14:50:50 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.26
+- Bump @fluentui-react-native/interactive-hooks to v0.22.1
+- Bump @fluentui-react-native/tokens to v0.20.0
+- Bump @fluentui-react-native/text to v0.19.0
+
+## 0.3.1
+
+Tue, 13 Dec 2022 18:37:10 GMT
+
+### Patches
+
+- Change disabled link to not be focusable (ruaraki@microsoft.com)
+
+## 0.3.0
+
+Fri, 09 Dec 2022 10:05:33 GMT
+
+### Minor changes
+
+- initial tokens for link android (rohanpd.work@gmail.com)
+
+## 0.2.7
+
+Fri, 09 Dec 2022 05:23:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.25
+- Bump @fluentui-react-native/interactive-hooks to v0.22.0
+- Bump @fluentui-react-native/tokens to v0.19.1
+- Bump @fluentui-react-native/text to v0.18.2
+
+## 0.2.6
+
+Tue, 06 Dec 2022 16:59:54 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.24
+- Bump @fluentui-react-native/interactive-hooks to v0.21.5
+- Bump @fluentui-react-native/text to v0.18.1
+
+## 0.2.5
+
+Tue, 06 Dec 2022 00:46:41 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.23
+- Bump @fluentui-react-native/interactive-hooks to v0.21.4
+- Bump @fluentui-react-native/tokens to v0.19.0
+- Bump @fluentui-react-native/text to v0.18.0
+
+## 0.2.4
+
+Thu, 01 Dec 2022 03:12:20 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.22
+- Bump @fluentui-react-native/interactive-hooks to v0.21.3
+- Bump @fluentui-react-native/tokens to v0.18.1
+- Bump @fluentui-react-native/text to v0.17.1
 
 ## 0.2.3